BAC Complete
Complete, tested and all is good with the BAC calculator. It will give
results for any form of alcohol, multiple types of alcohol and if the
gender is not m,M,f or F then the user will get an alert letting them
know that they need to input the correct data.